+++ packages/netglade_utils/test/src/extensions/iterable_extensions_test.dart
@@ -1,21 +1,8 @@
+import 'package:collection/collection.dart';
 import 'package:netglade_utils/netglade_utils.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('distinct', () {
-    test('distinct', () {
-      expect([1, 2, 3].distinct(), equals({1, 2, 3}));
-    });
-
-    test('with duplicates', () {
-      expect([2, 4, 4, 5, 6, 6, 6].distinct(), equals({2, 4, 5, 6}));
-    });
-
-    test('empty', () {
-      expect(<int>[].distinct(), equals(<int>{}));
-    });
-  });
-
   group('distinct by', () {
     test('distinct', () {
       expect(